fix frequency finder, broken by change to double prec for freqs

+++ so2sdr-bandmap/so2sdr-bandmap.h
@@ -154,8 +154,8 @@ private:
     QSlider              slider;
     QString              ipAddress;
     QTimer               tuningTimer;
-    int                  flow;
-    int                  fhigh;
+    double               flow;
+    double               fhigh;
     QTcpSocket           *socket;
     QTcpServer           server;
     QUdpSocket           socketUdp,socketUdpN1MM;
